all classes and files must belong to a named package, db.java qulice violation fixes
