+++ java-itu-minitwit/src/main/java/Logic/Hashing.java
@@ -1,3 +1,4 @@
+package Logic;
 
 import java.security.*;
 import java.nio.charset.StandardCharsets;
