+++ java-itu-minitwit/src/main/java/Logic/Logger.java
@@ -1,3 +1,5 @@
+package Logic;
+
 import java.lang.management.ManagementFactory;
 import java.lang.management.OperatingSystemMXBean;
 import java.util.HashMap;
