+++ java-itu-minitwit/src/main/java/Logic/Queries.java
@@ -1,3 +1,5 @@
+package Logic;
+
 import Model.Follower;
 import Model.Message;
 import Model.Tweet;
