+++ java-itu-minitwit/src/main/java/Logic/minitwit.java
@@ -1,3 +1,5 @@
+package Logic;
+
 import Model.Tweet;
 import Model.User;
 import RoP.Failure;
