+++ java-itu-minitwit/src/test/java/QueriesTest.java
@@ -1,4 +1,8 @@
 
+import Logic.DB;
+import Logic.Hashing;
+import Logic.Logger;
+import Logic.Queries;
 import RoP.Result;
 import RoP.Success;
 import org.junit.jupiter.api.Test;
